refactor(web): flatten examples — remove _components/ and _examples/
Co-locate each example in its route folder as example.tsx.
Move shared files (example-shell, example-page-shell) to examples root.
Inline registry data into gallery page. Delete no-op EditorThemeProvider.
