--- apps/web/src/app/editor/examples/_components/editor-theme-provider.tsx
@@ -1,9 +1,0 @@
-'use client';
-
-export function EditorThemeProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <>{children}</>;
-}
