+++ apps/web/src/app/editor/examples/basic-editor/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { BasicEditor as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { BasicEditor as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Basic Editor — Editor Examples',
