+++ apps/web/src/app/editor/examples/bubble-menu/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { BubbleMenuExample as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { BubbleMenuExample as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Bubble Menu — Editor Examples',
