+++ apps/web/src/app/editor/examples/buttons/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { Buttons as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { Buttons as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Buttons — Editor Examples',
