+++ apps/web/src/app/editor/examples/column-layouts/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { ColumnLayouts as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { ColumnLayouts as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Column Layouts — Editor Examples',
