+++ apps/web/src/app/editor/examples/custom-bubble-menu/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { CustomBubbleMenu as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { CustomBubbleMenu as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Custom Bubble Menu — Editor Examples',
