+++ apps/web/src/app/editor/examples/email-export/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { EmailExport as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { EmailExport as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Email Export — Editor Examples',
