+++ apps/web/src/app/editor/examples/full-email-builder/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { FullEmailBuilder as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { FullEmailBuilder as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Full Email Builder — Editor Examples',
