+++ apps/web/src/app/editor/examples/inspector-defaults/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { InspectorDefaults as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { InspectorDefaults as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Inspector — Defaults — Editor Examples',
