+++ apps/web/src/app/editor/examples/link-editing/example.tsx
@@ -3,7 +3,7 @@
 import { StarterKit } from '@react-email/editor/extensions';
 import { BubbleMenu } from '@react-email/editor/ui';
 import { EditorProvider } from '@tiptap/react';
-import { ExampleShell } from '../_components/example-shell';
+import { ExampleShell } from '../example-shell';
 
 const extensions = [StarterKit];
