+++ apps/web/src/app/editor/examples/one-line-editor/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { OneLineEditor as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { OneLineEditor as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'One-Line Editor — Minimal — Editor Examples',
