+++ apps/web/src/app/editor/examples/slash-commands/page.tsx
@@ -1,6 +1,6 @@
 import type { Metadata } from 'next';
-import { SlashCommands as Example } from '../_components/dynamic-examples';
-import { ExamplePageShell } from '../_components/example-page-shell';
+import { SlashCommands as Example } from '../dynamic-imports';
+import { ExamplePageShell } from '../example-page-shell';
 
 export const metadata: Metadata = {
   title: 'Slash Commands — Editor Examples',
